+++ tests/scanner_assets/client_rust_code.rs
@@ -86,6 +86,13 @@ pub mod wl_foo {
                 Request::CreateBar { .. } => 1,
             }
         }
+        fn since(&self) -> u32 {
+            match *self {
+                Request::__nonexhaustive => unreachable!(),
+                Request::FooIt { .. } => 1,
+                Request::CreateBar { .. } => 1,
+            }
+        }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
             match opcode {
                 1 => Some(Object::from_interface::<super::wl_bar::WlBar>(
@@ -149,6 +156,12 @@ pub mod wl_foo {
             match *self {
                 Event::__nonexhaustive => unreachable!(),
                 Event::Cake { .. } => 0,
+            }
+        }
+        fn since(&self) -> u32 {
+            match *self {
+                Event::__nonexhaustive => unreachable!(),
+                Event::Cake { .. } => 2,
             }
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
@@ -255,11 +268,11 @@ pub mod wl_foo {
         }
     }
     #[doc = r" The minimal object version supporting this request"]
-    pub const REQ_FOO_IT_SINCE: u16 = 1u16;
+    pub const REQ_FOO_IT_SINCE: u32 = 1u32;
     #[doc = r" The minimal object version supporting this request"]
-    pub const REQ_CREATE_BAR_SINCE: u16 = 1u16;
+    pub const REQ_CREATE_BAR_SINCE: u32 = 1u32;
     #[doc = r" The minimal object version supporting this event"]
-    pub const EVT_CAKE_SINCE: u16 = 2u16;
+    pub const EVT_CAKE_SINCE: u32 = 2u32;
 }
 #[doc = "Interface for bars\n\nThis interface allows you to bar your foos."]
 pub mod wl_bar {
@@ -335,6 +348,14 @@ pub mod wl_bar {
             match *self {
                 Request::__nonexhaustive => unreachable!(),
                 Request::BarDelivery { .. } => 0,
+                Request::Release => 1,
+                Request::_Self { .. } => 2,
+            }
+        }
+        fn since(&self) -> u32 {
+            match *self {
+                Request::__nonexhaustive => unreachable!(),
+                Request::BarDelivery { .. } => 2,
                 Request::Release => 1,
                 Request::_Self { .. } => 2,
             }
@@ -437,6 +458,12 @@ pub mod wl_bar {
             match *self {
                 Event::__nonexhaustive => unreachable!(),
                 Event::_Self { .. } => 0,
+            }
+        }
+        fn since(&self) -> u32 {
+            match *self {
+                Event::__nonexhaustive => unreachable!(),
+                Event::_Self { .. } => 2,
             }
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
@@ -625,13 +652,13 @@ pub mod wl_bar {
         }
     }
     #[doc = r" The minimal object version supporting this request"]
-    pub const REQ_BAR_DELIVERY_SINCE: u16 = 2u16;
+    pub const REQ_BAR_DELIVERY_SINCE: u32 = 2u32;
     #[doc = r" The minimal object version supporting this request"]
-    pub const REQ_RELEASE_SINCE: u16 = 1u16;
+    pub const REQ_RELEASE_SINCE: u32 = 1u32;
     #[doc = r" The minimal object version supporting this request"]
-    pub const REQ_SELF_SINCE: u16 = 2u16;
+    pub const REQ_SELF_SINCE: u32 = 2u32;
     #[doc = r" The minimal object version supporting this event"]
-    pub const EVT_SELF_SINCE: u16 = 2u16;
+    pub const EVT_SELF_SINCE: u32 = 2u32;
 }
 #[doc = "core global object\n\nThis global is special and should only generate code client-side, not server-side."]
 pub mod wl_display {
@@ -652,6 +679,11 @@ pub mod wl_display {
             }
         }
         fn opcode(&self) -> u16 {
+            match *self {
+                Request::__nonexhaustive => unreachable!(),
+            }
+        }
+        fn since(&self) -> u32 {
             match *self {
                 Request::__nonexhaustive => unreachable!(),
             }
@@ -683,6 +715,11 @@ pub mod wl_display {
             }
         }
         fn opcode(&self) -> u16 {
+            match *self {
+                Event::__nonexhaustive => unreachable!(),
+            }
+        }
+        fn since(&self) -> u32 {
             match *self {
                 Event::__nonexhaustive => unreachable!(),
             }
@@ -773,6 +810,12 @@ pub mod wl_registry {
                 Request::Bind { .. } => 0,
             }
         }
+        fn since(&self) -> u32 {
+            match *self {
+                Request::__nonexhaustive => unreachable!(),
+                Request::Bind { .. } => 1,
+            }
+        }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
             match opcode {
                 _ => None,
@@ -810,6 +853,11 @@ pub mod wl_registry {
             }
         }
         fn opcode(&self) -> u16 {
+            match *self {
+                Event::__nonexhaustive => unreachable!(),
+            }
+        }
+        fn since(&self) -> u32 {
             match *self {
                 Event::__nonexhaustive => unreachable!(),
             }
@@ -883,7 +931,7 @@ pub mod wl_registry {
         }
     }
     #[doc = r" The minimal object version supporting this request"]
-    pub const REQ_BIND_SINCE: u16 = 1u16;
+    pub const REQ_BIND_SINCE: u32 = 1u32;
 }
 #[doc = "callback object\n\nThis object has a special behavior regarding its destructor."]
 pub mod wl_callback {
@@ -904,6 +952,11 @@ pub mod wl_callback {
             }
         }
         fn opcode(&self) -> u16 {
+            match *self {
+                Request::__nonexhaustive => unreachable!(),
+            }
+        }
+        fn since(&self) -> u32 {
             match *self {
                 Request::__nonexhaustive => unreachable!(),
             }
@@ -945,6 +998,12 @@ pub mod wl_callback {
             match *self {
                 Event::__nonexhaustive => unreachable!(),
                 Event::Done { .. } => 0,
+            }
+        }
+        fn since(&self) -> u32 {
+            match *self {
+                Event::__nonexhaustive => unreachable!(),
+                Event::Done { .. } => 1,
             }
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
@@ -1015,5 +1074,5 @@ pub mod wl_callback {
         }
     }
     #[doc = r" The minimal object version supporting this event"]
-    pub const EVT_DONE_SINCE: u16 = 1u16;
+    pub const EVT_DONE_SINCE: u32 = 1u32;
 }